feat(runtime): WASM concurrency lowering — sequential default (phase-10 line 125)
spawn()/TaskGroup/par {} now work on wasm_wasi/wasm_browser via
cooperative sequential scheduling on the main thread. Before: spawn
programs failed at wasm link (native scheduler is net-gated, absent
from the wasm archive); par {} trapped at runtime (pool()'s
thread::spawn). Runtime-side — codegen's karac_* call surface is
unchanged; the wasm archive implements it sequentially:

- runtime/src/seq_scheduler.rs (new): FIFO ready queue — spawn
  enqueues (deferred, not inline), join points (task_join,
  taskgroup_join_and_free, defensively task_handle_free) drive the
  queue to the joined handle's terminal state. Full native export
  surface under cfg(all(wasm, not(net))) — collision-proof against
  scheduler.rs. drive_until_terminal is the documented hook for the
  line-129 yield-to-event-loop entry; spawn_coro deliberately absent
  (event-loop substrate). Deterministic spawn-order execution — a
  legal schedule of the native pool's racy one.
- seq_par_run (lib.rs): sequential karac_par_run body — source-order
  branches, per-call cancel flag + parent_cancel cascade + FrameGuard
  parity; pool substrate (ParCall/Task/Pool/workers/dispatch helpers)
  compiled out of the wasm archive. karac_par_reduce takes the
  single-worker fast path unconditionally on wasm.
- ABI fix surfaced in-slice: karac_par_run's count was usize vs
  codegen's all-targets i64 declaration — i32-width on wasm32, landing
  calls on a wasm-ld signature_mismatch stub (unreachable trap at the
  first par block; the __karac_malloc64 size_t class). Now u64
  (ABI-identical on 64-bit native); seq spawn export's size/align
  params likewise.
- Codegen: auto_par_disabled |= active_target_is_wasm() — fan-out is
  pure overhead on a single-threaded target; one gate covers parallel
  groups + reductions. Explicit par {} untouched (keeps karac_par_run
  dispatch for cancellation/result-slot semantics).
- Tests: runtime units x9 (seq scheduler x6, seq_par_run x3 — module
  compiles under native cfg(test)); tests/wasm_codegen.rs (new
  dedicated binary — set_active_target is process-global, wasm IR pins
  can't share a process with native codegen tests) x2; tests/cli.rs
  node:wasi E2E x2 (deterministic par branch order; deferred-spawn
  sentinel + FIFO worker order + h.join() result transport).
- Channel leg blocked, not dropped: channels have no AOT codegen on
  any target (interpreter-only, CR-202) — new phase-6 "Channel AOT
  codegen lowering" blocker entry, bidirectional cross-refs with both
  phase-10 entries. design.md § WASM Concurrency Lowering trued to
  shipped reality; CLAUDE.md + runtime/Cargo.toml archive notes
  updated.

Gates: fmt, clippy --all --all-targets --features llvm, wasm-target
clippy (-p karac-runtime --target wasm32-wasip1 --no-default-features),
full cargo test --features llvm (40/40 binaries green).

+++ runtime/src/scheduler.rs
@@ -1358,7 +1358,7 @@ mod tests {
                     let start = Instant::now();
                     for _ in 0..CALLS {
                         unsafe {
-                            karac_par_run(branches.as_ptr(), n, 0, std::ptr::null());
+                            karac_par_run(branches.as_ptr(), n as u64, 0, std::ptr::null());
                         }
                     }
                     start.elapsed()
@@ -1408,7 +1408,7 @@ mod tests {
                 let par_ns = median_ns(1, 9, || {
                     let start = Instant::now();
                     unsafe {
-                        karac_par_run(branches.as_ptr(), N, 0, std::ptr::null());
+                        karac_par_run(branches.as_ptr(), N as u64, 0, std::ptr::null());
                     }
                     start.elapsed()
                 });
